cleanup rest reading
