+++ src/robot/parsing/restreader.py
@@ -13,10 +13,49 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
+from robot.errors import DataError
 
-def read_rest(rstfile):
-    from .restsupport import publish_doctree, RobotDataStorage
+try:
+    from docutils.core import publish_doctree
+    from docutils.parsers.rst.directives import register_directive
+    from docutils.parsers.rst.directives.body import CodeBlock
+except ImportError:
+    raise DataError("Using reStructuredText test data requires having "
+                    "'docutils' module version 0.9 or newer installed.")
 
+
+class CaptureRobotData(CodeBlock):
+
+    def run(self):
+        if 'robotframework' in self.arguments:
+            store = RobotDataStorage(self.state_machine.document)
+            store.add_data(self.content)
+        return []
+
+
+register_directive('code', CaptureRobotData)
+register_directive('code-block', CaptureRobotData)
+register_directive('sourcecode', CaptureRobotData)
+
+
+class RobotDataStorage(object):
+
+    def __init__(self, doctree):
+        if not hasattr(doctree, '_robot_data'):
+            doctree._robot_data = []
+        self._robot_data = doctree._robot_data
+
+    def add_data(self, rows):
+        self._robot_data.extend(rows)
+
+    def get_data(self):
+        return '\n'.join(self._robot_data)
+
+    def has_data(self):
+        return bool(self._robot_data)
+
+
+def read_data(rstfile):
     doctree = publish_doctree(
         rstfile.read(), source_path=rstfile.name,
         settings_overrides={
@@ -25,4 +64,3 @@ def read_rest(rstfile):
         })
     store = RobotDataStorage(doctree)
     return store.get_data()
-
